requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0.post2
Flask-Bcrypt==1.0.1
Pillow==10.0.0
gunicorn==21.2.0
